+++ lib/browser/mixins/browser_wait.ts
@@ -5,7 +5,7 @@ import { WendigoSelector } from '../../types';
 import { createFindTextXPath, delay, isNumber } from '../../utils/utils';
 import FailIfNotLoaded from '../../decorators/fail_if_not_loaded';
 import OverrideError from '../../decorators/override_error';
-import { EvaluateFn } from '../../puppeteer_wrapper/puppeteer_types';
+import { EvaluateFn, PuppeteerErrors } from '../../puppeteer_wrapper/puppeteer_types';
 
 export default abstract class BrowserWait extends BrowserNavigation {
 
@@ -14,6 +14,7 @@ export default abstract class BrowserWait extends BrowserNavigation {
     }
 
     @FailIfNotLoaded
+    @OverrideError()
     public async waitFor(selector: EvaluateFn, timeout?: number, ...args: Array<any>): Promise<void> {
         timeout = this._getTimeout(timeout);
         args = args.map((e) => {
@@ -25,19 +26,20 @@ export default abstract class BrowserWait extends BrowserNavigation {
                 timeout: timeout,
                 visible: true
             }, ...args);
-        } catch (err) {
-            if (err instanceof Error && err.message.match(/DOMException\:/)) { // TODO: move to a helper/wrapper
-                throw new QueryError("waitFor", `Invalid selector "${selector}".`);
-            } else {
-                let errMsg;
-                if (typeof selector === 'function') errMsg = `Waiting for function to return true`;
-                else errMsg = `Waiting for element "${selector}"`;
-                throw new TimeoutError("waitFor", errMsg, timeout);
-            }
+        } catch (error) {
+            let errMsg;
+            if (typeof selector === 'function') errMsg = `Waiting for function to return true`;
+            else errMsg = `Waiting for element "${selector}"`;
+            throw this._overrideWaitError(error, {
+                timeoutMessage: errMsg,
+                selector: `${selector}`,
+                timeout
+            });
         }
     }
 
     @FailIfNotLoaded
+    @OverrideError()
     public async waitUntilNotVisible(selector: WendigoSelector, timeout?: number): Promise<void> {
         timeout = this._getTimeout(timeout);
         try {
@@ -46,11 +48,16 @@ export default abstract class BrowserWait extends BrowserNavigation {
                 return !WendigoUtils.isVisible(element);
             }, timeout, selector);
         } catch (err) {
-            throw new TimeoutError("waitUntilNotVisible", `Waiting for element "${selector}" to not be visible`, timeout);
+            throw this._overrideWaitError(err, {
+                timeoutMessage: `Waiting for element "${selector}" to not be visible`,
+                selector: `${selector}`,
+                timeout
+            });
         }
     }
 
     @FailIfNotLoaded
+    @OverrideError()
     public async waitForUrl(url: string | RegExp, timeout?: number): Promise<void> {
         timeout = this._getTimeout(timeout);
         if (!url) return Promise.reject(new WendigoError("waitForUrl", `Invalid parameter url.`));
@@ -73,11 +80,15 @@ export default abstract class BrowserWait extends BrowserNavigation {
                 }
             }, timeout, parsedUrl);
         } catch (err) {
-            throw new TimeoutError("waitForUrl", `Waiting for url "${url}"`, timeout);
+            throw this._overrideWaitError(err, {
+                timeoutMessage: `Waiting for url "${url}"`,
+                timeout
+            });
         }
     }
 
     @FailIfNotLoaded
+    @OverrideError()
     public async waitForNavigation(timeout?: number): Promise<void> {
         timeout = this._getTimeout(timeout);
         const t1 = new Date().getTime();
@@ -94,7 +105,10 @@ export default abstract class BrowserWait extends BrowserNavigation {
                 return Boolean(w.WendigoUtils);
             }, timeout2);
         } catch (err) {
-            throw new TimeoutError("waitForNavigation", "", timeout);
+            throw this._overrideWaitError(err, {
+                timeoutMessage: "",
+                timeout
+            });
         }
     }
 
@@ -109,6 +123,8 @@ export default abstract class BrowserWait extends BrowserNavigation {
         return result[1];
     }
 
+    @FailIfNotLoaded
+    @OverrideError()
     public async waitForText(text: string, timeout?: number): Promise<void> {
         timeout = this._getTimeout(timeout);
         try {
@@ -117,21 +133,30 @@ export default abstract class BrowserWait extends BrowserNavigation {
                 return Boolean(WendigoUtils.xPathQuery(xp).length > 0);
             }, timeout, xPath);
         } catch (err) {
-            throw new TimeoutError("waitForText", `Waiting for text "${text}"`, timeout);
+            throw this._overrideWaitError(err, {
+                timeoutMessage: `Waiting for text "${text}"`,
+                timeout
+            });
         }
     }
 
+    @FailIfNotLoaded
+    @OverrideError()
     public async waitAndClick(selector: string, timeout?: number): Promise<number> {
         timeout = this._getTimeout(timeout);
         try {
             await this.waitFor(selector, timeout);
             return await this.click(selector);
         } catch (err) {
-            throw new TimeoutError("waitAndClick", "", timeout);
+            throw this._overrideWaitError(err, {
+                timeoutMessage: "",
+                timeout
+            });
         }
     }
 
     @OverrideError()
+    @FailIfNotLoaded
     public async waitAndType(selector: string, text: string, timeout?: number): Promise<void> {
         await this.waitFor(selector, timeout);
         return await this.type(selector, text);
@@ -139,11 +164,14 @@ export default abstract class BrowserWait extends BrowserNavigation {
     }
 
     @OverrideError()
+    @FailIfNotLoaded
     public async waitAndCheck(selector: string, timeout?: number): Promise<void> {
         await this.waitFor(selector, timeout);
         return await this.check(selector);
     }
 
+    @OverrideError()
+    @FailIfNotLoaded
     public async waitUntilEnabled(selector: WendigoSelector, timeout?: number): Promise<void> {
         timeout = this._getTimeout(timeout);
         try {
@@ -154,12 +182,25 @@ export default abstract class BrowserWait extends BrowserNavigation {
                 return value === null;
             }, timeout, selector);
         } catch (err) {
-            throw new TimeoutError("waitUntilEnabled", `Waiting for element "${selector}" to be enabled`, timeout);
+          throw this._overrideWaitError(err, {
+              timeoutMessage: `Waiting for element "${selector}" to be enabled`,
+              timeout
+          });
         }
     }
 
     private _getTimeout(timeout?: number): number {
         if (isNumber(timeout)) return timeout;
         else return this._settings.defaultTimeout;
+    }
+
+    // TODO: this need to be completely reworked to make a sensible approach to error handling
+    private _overrideWaitError(error: Error, options: { timeoutMessage: string, timeout: number, selector?: string }): Error {
+        if (error instanceof PuppeteerErrors.TimeoutError || error instanceof TimeoutError) {
+            return new TimeoutError("", options.timeoutMessage, options.timeout);
+        }
+        if (error instanceof Error && error.message.match(/DOMException\:/) || error instanceof QueryError) {
+            return new QueryError("", `Invalid selector "${options.selector || ""}".`);
+        } else return error;
     }
 }